pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdb2reaction"
description = "pdb2reaction"
readme = "README.md"
authors = [{name = "Takuto Ohmura"}]
license = {file = "LICENSE"}
dynamic = ["version"]
requires-python = ">=3.11"
dependencies = [
    "fairchem-core",
    "pysisyphus @ git+https://github.com/t-0hmura/torch_pysis.git", 
    "torch>=2.6.0",
    "torch_geometric",
    "torch_scatter",
    "numpy",
    "biopython",
    "ase>=3.25.0",
    "pyyaml",
    "click"
]

[project.scripts]
pdb2reaction = "pdb2reaction.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["pdb2reaction*"]

[tool.setuptools]
package-data = { "pdb2reaction" = ["py.typed"] }

[tool.setuptools.dynamic]
version = {attr = "pdb2reaction.__version__"}
